Patch_CollisionLine
Update
1. 캐릭터 클래스 생성 및 GameMain의 캐릭터 조작 관련 코드 모두 캐릭터에서 구현하게끔 변경
2. CollisionBox 클래스 생성 및 ModelPart에서 구한 Min/Max 값을 받아와 CollisionLine 구현(현재 y축 기준 원 구현)

Issue
1. 캐릭터 회전 후 부들거림
2. CollisionLine이 반원으로 출력됨(index가 문제인지 확인 필요)
3. CollisionBox 클래스에서 min/max 값 받아올 때 캐릭터의 원본scale 값으로 받아옴(줄인 값을 받아와야함)
4. 회전각도별 이동방향 변경 필요

+++ Model/ModelAnimation.cpp
@@ -2,7 +2,7 @@
 #include "ModelAnimation.h"
 
 ModelAnimation::ModelAnimation(wstring name, int keyFrame, float defaultFrameRate)
-	: name(name), keyFrame(keyFrame), frameRate(keyFrame), defaultFrameRate(defaultFrameRate)
+	: name(name), keyFrame(keyFrame), frameRate((float)keyFrame), defaultFrameRate(defaultFrameRate)
 {
 	
 }
